requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
streamlit
pandas
plotly
scikit-learn
requests
